Added pause, stop and step functionality.
Added pause, stop and step functionality for neural network execution.
This was done by restructuring the way the neural network was being
executed. Now the neural network first computes the solution for it's
input sets and stores the meta-outputs. These outputs are then used to
animate the neural network in a step wise manner.
Also moved training options to a ribbon tab of its own.